Add protections so only admins can view mapreduce status pages.
Summary:
We're getting rid of `login: admin`, since appengine flex doesn't
support it, and moving the relevant logic into code.  Much of the
mapreduce library was already protected; in particular most methods
can only be run from a taskqueue.

But the UI code was only protected by `login: admin`.  This change
moves the admin check to in-code.  All UI goes through the same base
handler, so this was a very simple change.

Test Plan:
I removed `login: admin` from `/mapreduce(/.*)?` in
handlers-webapp.yaml, and then ran `make serve`.

I visited
   http://localhost:8080/mapreduce/status
as a logged out user and it said: "Could not parse response JSON
data," which is because the JSON route is off limits.  (The frame of
the webpage still displays because I didn't see a point in blocking
the static content).

I then redid the experiment, with the same results, as a logged in
user who was not an admin.

I then did it again with a logged in user who *was* an admin.  In this
case, the page rendered properly, and I was presented with the "Launch
job" dropdown.

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D34407